refactor(geometry): split kernel/arrangement.rs into mod + classify + boolean + tests
arrangement.rs (1489 LOC), the determinism firewall, -> arrangement/{mod (types
+ arrange/arrange_many, 383), classify (point-inside + surface-detection + the
3-regime boolean_vids, 641), boolean (public drivers + boolean_manifest, 221),
tests (277)}.

Pure code motion (byte-exact sed extraction; each body diffed identical vs HEAD).
~11 helpers raised to pub(super) for cross-submodule reach; public API stays pub
and is re-exported from mod.rs so crate::kernel::arrangement::* paths are
identical; kernel/mod.rs untouched. classify.rs (641) kept whole as the cohesive
3-regime classification algorithm (interlocking determinism invariants).

DETERMINISM VERIFIED: boolean_manifest_is_pinned passes - hash 0x0465b83a5fdb8b2b
unchanged, native==wasm bit-identical. build/clippy -D warnings clean.
